CollectionView get data for item node
